Dashboard db backend

+++ controllers/api/user-routes.js
@@ -4,20 +4,28 @@ const authenticate = require('../../util/authenticate');
 
 // `/api/user` HTML request endpoint
 
-// Get all users, endpoint=/api/user/
+// Get all users, admin use only
 router.get('/', (req, res) => {
-  User.findAll({
-      attributes: {exclude: ['password']}
-  })
-  .then(dbPostData => res.json(dbPostData))
-  .catch(err => {
-      console.log(err);
-      res.status(500).json(err);
+    if (!req.session.loggedIn || req.session.id != 1) {
+        res.redirect('/');
+        return;
+    }
+    User.findAll({
+          attributes: {exclude: ['password']}
+    })
+    .then(dbPostData => res.json(dbPostData))
+    .catch(err => {
+          console.log(err);
+          res.status(500).json(err);
     });
 });
 
-// Get user by id, endpoint=/api/user/:id/
+// Get user by id, admin use only
 router.get('/:id', (req, res) => {
+    if (!req.session.loggedIn || req.session.id != 1) {
+        res.redirect('/');
+        return;
+    }
     User.findOne({
         where: { id: req.params.id },
         attributes: { exclude: ['password', 'id'] }
@@ -31,8 +39,12 @@ router.get('/:id', (req, res) => {
 
 // Delete a user by id
 router.delete('/:id', authenticate, (req, res) => {
+    if (!req.session.loggedIn) {
+        res.redirect('/');
+        return;
+    }
     User.destroy({
-        where: { id: req.params.id },
+        where: { id: req.params.user_id },
     })
     .then(dbPostData => res.json(dbPostData))
     .catch(err => {
